Fix graph height and improve Y axes label
